requirements.txt
PyNaCl>=1.5.0
aiohttp>=3.9.0
aioconsole>=0.7.0
